+++ excalidraw-assets/src/index.tsx
@@ -1,12 +1,11 @@
 import React from "react";
-import { createRoot } from "react-dom/client"; // From React 18
+import { createRoot } from "react-dom/client";
 
-// import "./styles.scss";
 import "./styles.css";
 
-import {App} from "./App";
+import { App } from "./App";
 
-// https://reactjs.org/blog/2022/03/08/react-18-upgrade-guide.html#updates-to-client-rendering-apis
+// https://react.dev/blog/2024/04/25/react-19-upgrade-guide
 const rootContainer = document.getElementById("root");
-const root = createRoot(rootContainer!)
+const root = createRoot(rootContainer!);
 root.render(React.createElement(App));